asynq: don't catch BaseExceptions
Otherwise, exceptions like SystemExit can get swallowed for a while
during the execution of other tasks, which is undesirable behavior.

However, I am not sure this change will not lead to corruption of
the internal state of the scheduler in some cases. We'll need to
think more about whether that is safe.